Add tabs view; enhance the gradient editor
* Now uses tabs for palettes and custom gradients
* Allow reorder the colors in the custom gradient view
* Style enhancements
