+++ sass/components/components.scss
@@ -10,6 +10,7 @@
 @import "color_picker.scss";
 @import "file_view.scss";
 @import "image_uploader.scss";
+@import "tabs_view.scss";
 
 .charticulator__error-boundary-report {
   p {
