+++ src/app/components/index.ts
@@ -7,6 +7,7 @@ export * from "./selecting";
 export * from "./split_panes";
 export * from "./zoomable";
 export * from "./color_picker";
+export * from "./gradient_picker";
 export * from "./minimizable_panel";
 export * from "./error_boundary";
 export * from "./dropdown";
