+++ src/tests/test_app/color_picker.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 import * as React from "react";
-import { ColorPicker } from "../../app/components";
+import { ColorPicker, GradientPicker } from "../../app/components";
 
 export class ColorPickerTestView extends React.Component<{}, {}> {
   public render() {
@@ -27,6 +27,26 @@ export class ColorPickerTestView extends React.Component<{}, {}> {
   }
 }
 
+export class GradientPickerTestView extends React.Component<{}, {}> {
+  public render() {
+    return (
+      <div>
+        <div
+          style={{
+            background: "#eee",
+            border: "10px solid #aaa",
+            width: "300px",
+            display: "inline-block"
+          }}
+        >
+          <GradientPicker />
+        </div>
+      </div>
+    );
+  }
+}
+
 export function register(f: any) {
   f("ColorPicker", ColorPickerTestView);
+  f("GradientPicker", GradientPickerTestView);
 }
